project/requirements.txt
Flask==2.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
werkzeug==2.2.3
opencv-python-headless==4.5.5.62
pytesseract==0.3.9
spacy==3.4.1
hunspell==0.5.5
python-dateutil==2.8.2
Pillow==9.5.0